requirements.txt
streamlit
langgraph>
langchain-core>
langchain-google-genai
python-dotenv>
uuid
